more post revision issues
git-svn-id: file:///Users/michaelkoepke/svn/semiologic/plugins/script-manager/trunk@159 6c9c31c4-cb3d-4424-af3d-876279169c7e

+++ script-manager-admin.php
@@ -232,6 +232,10 @@ EOF;
 
 	function save_entry($post_ID)
 	{
+		$post = get_post($post_ID);
+		
+		if ( $post->post_type == 'revision' ) return;
+		
 		if ( current_user_can('unfiltered_html') )
 		{
 			delete_post_meta($post_ID, '_scripts_override');
